Macrostrat columns (#116)
* Adding `from_macrostrat` to striplog to generate chronostratigraphic columns for a given location.

This adds a `from_macrostrat` method to `striplog.striplog`. It expects a lng lat point, and will then use the MacroStrat API to get nearby geological units and construct a column based on the ages of the rocks within a buffer zone of the point (by default 0.2 degrees, but easily changed through a parameter). The column is returned as a `striplog.Striplog`.

While in `striplog.py`, this is not part of the `Striplog` class.

It is possible that I should just refactor this so that it does work as part of the class, but there are a handful of additional functions that handle part of the data munging. These could be pulled into the main function which could be moved into the class proper, if so desired. They have relatively little utility outside this specific application.

* Return legend with striplog