Working on tests
Starting to whittle down on the tests that need to pass.

One thing that I am doing a fair bit that feels off is having to duplicate data internally, with both `my_interval.<thing>` and `my_interval.data[<thing>]` existing for some things. It would be best to only have one, but the `self.data` is something that we have as a result of extending `UserDict`. I think that we can avoid that with a `MutableMapping`, but I would need to look into it more.