Pass most tests
Some fairly major changes:
`self.components`, `self.top`, `self.base` and `self.decription` are used commonly. To prevent duplication, these are now defined as properties. This means that most of the existing tests and code works (although I have not reverted all of my changes, so there are still calls that have been changed to `self.data.get('top')`, which can probably just be `self.top` now.

All of the Interval tests now pass, although doing things of the form `my_interval.top = 25` does not seem to work any more. This is obviously a breaking change, but it matches how dictionaries and similar things (pandas DataFrames and such) work.

This last needs a change in the tests, because the tests use `x[-1].base = 65`, which now needs to be `x[-1]['base'] = 65`. This is the only change to a test that I have made some far, although I expect that coverage has been reduced.

`my_interval.data` will return everything, which might not be what we want to do? I am looking at how to get a different `data` dictionary that contains only things that are not already used, which is the current behaviour for things like `_repr_html_()`.