Passing further tests
Pass tests relating to `Striplog`s using the new `Interval`. These required changes akin to the others, calling something like `my_interval['top'] = 10` instead of `my_interval.top = 10`, as previously detailed.

There is also a needed change in the `Interval.union` method. Most things can be converted to lists and concatenated, but we are already dealing with the new `top` and `base`, so those are now ignored in that method.