All tests passing
All tests are now passing locally.

We do define a `base` in an `Interval` if we are not given one (by making it the same as `top`).

As of now, `Interval`s are `dict`s with some extra methods. This can possibly be done on `Striplog`s proper now as well. Note that some of the docs need to be rewritten to accommodate these changes, principally the inability to use `Interval.top = value` and similar, so this is a breaking change that should be introduced over a couple of releases.