Removed overcomplicated new code
Because the `Interval.top` and similar things are working, some of the code that I wrote to handle accessing those is already obsolete. This strips is back to something akin to how it originally looked, since the API has not really been altered as much as I thought it was going to.