Modernise the build/packaging process
Striplog is a relatively old library, and the way that it is built is not the same as our more recent efforts. This branch (and commits to it) are intended to make striplog consistent with those changes prior to a rebuilding of the documentation to also match more recent libraries.