Edit author and description in `setup.cfg`
Was using the wrong stuff here. `AUTHORS.md` will be written as part of the docs update if it is not already. The URL will currently not work, again this is part of the docs changes on another branch.

Also removed the `_version.py` file, since that is now autogenerated.