Move the tutorials into the `docs` folder
Sphinx does not really like looking outside the docs folder, so this is a much easier way of working around that limitation.

Also need to point the `README`'s link to the new location. There are possibly some changes that need to be made to the notebooks, but those will follow.