Fix #135 (#136)
Fixes #135
The Striplog constructor will now create Components if there is misc data given to it. If there are not components, then the `description` and either the provided `Lexicon` or the default one will be passed to the `Interval` constructor.

There is a new test for this case as well.