IO changes (#134)
* Update striplog.py

Simplifying and extending `from_csv`.

Initial test uses `np.genfromtext` internally. Explicitly expects to be told which columns the `top`, `base` and/or `thickness` are in. Other columns will be added to a dictionary from which the intervals will be created.

Addresses, but does not entirely close #128

* Update `from_csv`

Expanded docs. The will need to be restyled to match the rest of the library, but this will work for now.
Added a series of FutureWarnings for args that we want to remove going forward.
Added some new args.

* Updated `from_csv`

Updated docs and args to be the same.
Changed version for deprecation to 0.9.1.

* Updated test suite

Changed tests to use new `from_csv` method.

The tests were passing in strings. These no need to be explicitly converted to `io.StringIO` objects, along with the `names=True`.

* Updated `from_csv()`

Removed extra args. These are not currently being used, but we might want them later. No need for them for now.

* Remove old `from_csv`

Removed old `from_csv` function - it is no longer being used anyway.